+++ propulsion-summary-projector/Program.fs
@@ -36,7 +36,7 @@ module CmdParser =
     open Equinox.EventStore
     [<NoEquality; NoComparison>]
     type Parameters =
-        | [<MainCommand; ExactlyOnce>]      ConsumerGroupName of string
+        | [<AltCommandLine "-g"; Mandatory>] ConsumerGroupName of string
         | [<AltCommandLine "-r"; Unique>]   MaxReadAhead of int
         | [<AltCommandLine "-w"; Unique>]   MaxWriters of int
         | [<AltCommandLine "-v"; Unique>]   Verbose